Function libm::asinf

pub fn asinf(x: f32) -> f32

Arcsine (f32)

Computes the inverse sine (arc sine) of the argument x. Arguments to asin must be in the range -1 to 1. Returns values in radians, in the range of -pi/2 to pi/2.