Function libm::fmaf

pub fn fmaf(x: f32, y: f32, z: f32) -> f32

Floating multiply add (f32)

Computes (x*y)+z, rounded as one ternary operation: Computes the value (as if) to infinite precision and rounds once to the result format, according to the rounding mode characterized by the value of FLT_ROUNDS.