libm::math::support::env

Struct Status

pub struct Status(u8);

IEEE 754 exception status flags.

Tuple Fields
0: u8
Implementations

impl Status

pub const OK: Self

Default status indicating no errors.

pub const INVALID: Self

No definable result.

Includes:

Any ops on sNaN, with a few exceptions.

0 * inf, inf * 0.

fma(0, inf, c) or fma(inf, 0, c), possibly excluding c = qNaN.

+inf + -inf and similar (includes subtraction and fma).

0.0 / 0.0, inf / inf

remainder(x, y) if y == 0.0 or x == inf, and neither is NaN.

sqrt(x) with x < 0.0.

pub const DIVIDE_BY_ZERO: Self

Division by zero.

The default result for division is +/-inf based on operand sign. For logB, the default result is -inf. x / y when x != 0.0 and y == 0.0,

pub const OVERFLOW: Self

The result exceeds the maximum finite value.

The default result depends on rounding mode. Nearest* rounds to +/- infinity, sign based on the intermediate result. Zero rounds to the signed maximum finite. Positive and Negative round to signed maximum finite in one direction, signed infinity in the other.

pub const UNDERFLOW: Self

The result is subnormal and lost precision.

pub const INEXACT: Self

The finite-precision result does not match that of infinite precision, and the reason is not represented by one of the other flags.

pub const fn underflow(self) -> bool

True if UNDERFLOW is set.

pub const fn overflow(self) -> bool

True if OVERFLOW is set.

pub fn set_underflow(&mut self, val: bool)

pub const fn inexact(self) -> bool

True if INEXACT is set.

pub fn set_inexact(&mut self, val: bool)

fn set_flag(&mut self, val: bool, mask: Self)

pub(crate) const fn with(self, rhs: Self) -> Self