libm::math::support::float_traits

Trait HFloat

pub trait HFloat: Float {
    type D: DFloat<H = Self>;

    // Required method
    fn widen(self) -> Self::D;
}

Trait for floats half the bit width of another float.

Required Associated Types

type D: DFloat<H = Self>

Float that is double the bit width of the float this trait is implemented for.

Required Methods

fn widen(self) -> Self::D

Widen the float type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.