libm::math::support::int_traits

Trait CastInto

pub trait CastInto<T: Copy>: Copy {
    // Required methods
    fn cast(self) -> T;
    fn cast_lossy(self) -> T;
}

Trait to express (possibly lossy) casting of integers

Required Methods

fn cast(self) -> T

By default, casts should be exact.

fn cast_lossy(self) -> T

Call for casts that are expected to truncate.

In practice, this is exactly the same as cast; the main difference is to document intent in code. cast may panic in debug mode.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.