Crate libm

libm in pure Rust

Structs

Libm

Generic helper for libm functions, abstracting over f32 and f64.

Functions

acos

Arccosine (f64)

acosf

Arccosine (f32)

acosh

Inverse hyperbolic cosine (f64)

acoshf

Inverse hyperbolic cosine (f32)

asin

Arcsine (f64)

asinf

Arcsine (f32)

asinh

Inverse hyperbolic sine (f64)

asinhf

Inverse hyperbolic sine (f32)

atan

Arctangent (f64)

atan2

Arctangent of y/x (f64)

atan2f

Arctangent of y/x (f32)

atanf

Arctangent (f32)

atanh

Inverse hyperbolic tangent (f64)

atanhf

Inverse hyperbolic tangent (f32)

cbrt

Compute the cube root of the argument.

cbrtf

Cube root (f32)

ceil

Ceil (f64)

ceilf

Ceil (f32)

copysign

Sign of Y, magnitude of X (f64)

copysignf

Sign of Y, magnitude of X (f32)

cos

The cosine of x (f64).

cosf

The cosine of x (f32).

cosh

Hyperbolic cosine (f64)

coshf

Hyperbolic cosine (f64)

erf

Error function (f64)

erfc

Complementary error function (f64)

erfcf

Complementary error function (f32)

erff

Error function (f32)

exp

Exponential, base e (f64)

exp2

Exponential, base 2 (f64)

exp2f

Exponential, base 2 (f32)

exp10

Calculates 10 raised to the power of x (f64).

exp10f

Calculates 10 raised to the power of x (f32).

expf

Exponential, base e (f32)

expm1

Exponential, base e, of x-1 (f64)

expm1f

Exponential, base e, of x-1 (f32)

fabs

Absolute value (magnitude) (f64)

fabsf

Absolute value (magnitude) (f32)

fdim

Positive difference (f64)

fdimf

Positive difference (f32)

floor

Floor (f64)

floorf

Floor (f32)

fma

Fused multiply add (f64)

fmaf

Floating multiply add (f32)

fmax

Return the greater of two arguments or, if either argument is NaN, the other argument.

fmaxf

Return the greater of two arguments or, if either argument is NaN, the other argument.

fmaximum

Return the greater of two arguments or, if either argument is NaN, the other argument.

fmaximum_num

Return the greater of two arguments or, if either argument is NaN, NaN.

fmaximum_numf

Return the greater of two arguments or, if either argument is NaN, NaN.

fmaximumf

Return the greater of two arguments or, if either argument is NaN, the other argument.

fmin

Return the lesser of two arguments or, if either argument is NaN, the other argument.

fminf

Return the lesser of two arguments or, if either argument is NaN, the other argument.

fminimum

Return the lesser of two arguments or, if either argument is NaN, the other argument.

fminimum_num

Return the lesser of two arguments or, if either argument is NaN, NaN.

fminimum_numf

Return the lesser of two arguments or, if either argument is NaN, NaN.

fminimumf

Return the lesser of two arguments or, if either argument is NaN, the other argument.

fmod

Calculate the remainder of x / y, the precise result of x - trunc(x / y) * y.

fmodf

Calculate the remainder of x / y, the precise result of x - trunc(x / y) * y.

frexp

frexpf

hypot

hypotf

ilogb

ilogbf

j0

Zeroth order of the Bessel function of the first kind (f64).

j0f

Zeroth order of the Bessel function of the first kind (f32).

j1

First order of the Bessel function of the first kind (f64).

j1f

First order of the Bessel function of the first kind (f32).

jn

Integer order of the Bessel function of the first kind (f64).

jnf

Integer order of the Bessel function of the first kind (f32).

ldexp

ldexpf

lgamma

The natural logarithm of the Gamma function (f64).

lgamma_r

lgammaf

The natural logarithm of the Gamma function (f32).

lgammaf_r

log

The natural logarithm of x (f64).

log2

The base 2 logarithm of x (f64).

log1p

The natural logarithm of 1+x (f64).

log1pf

The natural logarithm of 1+x (f32).

log2f

The base 2 logarithm of x (f32).

log10

The base 10 logarithm of x (f64).

log10f

The base 10 logarithm of x (f32).

logf

The natural logarithm of x (f32).

modf

modff

nextafter

nextafterf

pow

Returns x to the power of y (f64).

powf

Returns x to the power of y (f32).

remainder

remainderf

remquo

remquof

rint

Round x to the nearest integer, breaking ties toward even.

rintf

Round x to the nearest integer, breaking ties toward even.

round

Round x to the nearest integer, breaking ties away from zero.

roundeven

Round x to the nearest integer, breaking ties toward even. This is IEEE 754 roundToIntegralTiesToEven.

roundevenf

Round x to the nearest integer, breaking ties toward even. This is IEEE 754 roundToIntegralTiesToEven.

roundf

Round x to the nearest integer, breaking ties away from zero.

scalbn

scalbnf

sin

The sine of x (f64).

sincos

Both the sine and cosine of x (f64).

sincosf

Both the sine and cosine of x (f32).

sinf

The sine of x (f32).

sinh

The hyperbolic sine of x (f64).

sinhf

The hyperbolic sine of x (f32).

sqrt

The square root of x (f64).

sqrtf

The square root of x (f32).

tan

The tangent of x (f64).

tanf

The tangent of x (f32).

tanh

The hyperbolic tangent of x (f64).

tanhf

The hyperbolic tangent of x (f32).

tgamma

The Gamma function (f64).

tgammaf

The Gamma function (f32).

trunc

Rounds the number toward 0 to the closest integral value (f64).

truncf

Rounds the number toward 0 to the closest integral value (f32).

y0

Zeroth order of the Bessel function of the second kind (f64).

y0f

Zeroth order of the Bessel function of the second kind (f32).

y1

First order of the Bessel function of the second kind (f64).

y1f

First order of the Bessel function of the second kind (f32).

yn

Integer order of the Bessel function of the second kind (f64).

ynf

Integer order of the Bessel function of the second kind (f32).