libm::math

Module arch

Architecture-specific routines and operations.

LLVM will already optimize calls to some of these in cases that there are hardware instructions. Providing an implementation here just ensures that the faster implementation is used when calling the function directly. This helps anyone who uses libm directly, as well as improving things when these routines are called as part of other implementations.

Re-exports

pub use x86::sqrt;

pub use x86::sqrtf;

pub use x86::fma;

pub use x86::fmaf;

Modules

x86 🔒

Architecture-specific support for x86-32 and x86-64 with SSE2