libm::math::generic::sqrt

Trait SqrtHelper

pub trait SqrtHelper: Float {
    type ISet1: HInt + Into<Self::ISet2> + CastFrom<Self::Int> + From<u8>;
    type ISet2: HInt + From<Self::ISet1> + From<u8>;

    const FINAL_ROUNDS: u32;
    const SET1_ROUNDS: u32 = 0u32;
    const SET2_ROUNDS: u32 = 0u32;
}

Size-specific constants related to the square root routine.

Required Associated Constants

const FINAL_ROUNDS: u32

Number of rounds at Self::Int.

Provided Associated Constants

const SET1_ROUNDS: u32 = 0u32

Number of rounds at ISet1.

const SET2_ROUNDS: u32 = 0u32

Number of rounds at ISet2.

Required Associated Types

type ISet1: HInt + Into<Self::ISet2> + CastFrom<Self::Int> + From<u8>

Integer for the first set of rounds. If unused, set to the same type as the next set.

type ISet2: HInt + From<Self::ISet1> + From<u8>

Integer for the second set of rounds. If unused, set to the same type as the next set.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.