libm::math

Module j1

Constants

INVSQRTPI ğŸ”’

PR2 ğŸ”’

PR3 ğŸ”’

PR5 ğŸ”’

PR8 ğŸ”’

PS2 ğŸ”’

PS3 ğŸ”’

PS5 ğŸ”’

PS8 ğŸ”’

QR2 ğŸ”’

QR3 ğŸ”’

QR5 ğŸ”’

QR8 ğŸ”’

QS2 ğŸ”’

QS3 ğŸ”’

QS5 ğŸ”’

QS8 ğŸ”’

R00 ğŸ”’

R01 ğŸ”’

R02 ğŸ”’

R03 ğŸ”’

S01 ğŸ”’

S02 ğŸ”’

S03 ğŸ”’

S04 ğŸ”’

S05 ğŸ”’

TPI ğŸ”’

U0 ğŸ”’

V0 ğŸ”’

Functions

common ğŸ”’

j1

First order of the Bessel function of the first kind (f64).

pone ğŸ”’

qone ğŸ”’

y1

First order of the Bessel function of the second kind (f64).