libm::math

Module support

Re-exports

pub use big::i256;

pub use big::u256;

pub use env::FpResult;

pub use env::Round;

pub use env::Status;

pub use float_traits::DFloat;

pub use float_traits::Float;

pub use float_traits::HFloat;

pub use float_traits::IntTy;

pub use hex_float::Hexf;

pub use hex_float::hf32;

pub use hex_float::hf64;

pub use int_traits::CastFrom;

pub use int_traits::CastInto;

pub use int_traits::DInt;

pub use int_traits::HInt;

pub use int_traits::Int;

pub use int_traits::MinInt;

Modules

big ğŸ”’

Integers used for wide operations, larger than u128.

env ğŸ”’

Support for rounding directions and status flags as specified by IEEE 754.

feature_detect ğŸ”’

Helpers for runtime target feature detection that are shared across architectures.

float_traits ğŸ”’

hex_float

Utilities for working with hex float formats.

int_traits ğŸ”’

macros

Functions

cold_path

Hint to the compiler that the current path is cold.