libm::math::support

Module float_traits

Macros

float_impl 🔒

impl_d_float 🔒

impl_h_float 🔒

Traits

DFloat

Trait for floats twice the bit width of another integer.

Float

Trait for some basic operations on floats

HFloat

Trait for floats half the bit width of another float.

Functions

f32_from_bits

f32::from_bits

f32_to_bits

f32::to_bits

f64_from_bits

f64::from_bits

f64_to_bits

f64::to_bits

Type Aliases

IntTy

Access the associated Int type from a float (helper to avoid ambiguous associated types).