libm::math::support

Module hex_float

Utilities for working with hex float formats.

Structs

HexFloatParseError

Hexf

Format a floating point number as its IEEE hex (%a) representation.

Parsed 🔒

A parsed finite and unsigned floating point number.

Functions

dec_digit 🔒

fmt_any_hex 🔒

hex_digit 🔒

hf32

Construct a 32-bit float from hex float representation (C-style)

hf64

Construct a 64-bit float from hex float representation (C-style)

parse_any

Parse any float from hex to its bitwise representation.

parse_finite 🔒

parse_hex 🔒

Parse a hexadecimal float x

parse_hex_exact

Parses any float to its bitwise representation, returning an error if it cannot be represented exactly

shr2_round 🔒

Divide by 4, rounding with the given mode

shr_odd_rounding 🔒

Shift right, rounding all inexact divisions to the nearest odd number E.g. (0 >> 4) -> 0, (1..=31 >> 4) -> 1, (32 >> 4) -> 2, …

u128_ilog2 🔒

u128::ilog2