libm::math::support

Module int_traits

Macros

cast_into 🔒

cast_into_float 🔒

impl_d_int 🔒

impl_h_int 🔒

int_impl 🔒

int_impl_common 🔒

Traits

CastFrom

CastInto

Trait to express (possibly lossy) casting of integers

DInt

Trait for integers twice the bit width of another integer. This is implemented for all primitives except for u8, because there is not a smaller primitive.

HInt

Trait for integers half the bit width of another integer. This is implemented for all primitives except for u128, because it there is not a larger primitive.

Int

Trait for some basic operations on integers

MinInt

Minimal integer implementations needed on all integer types, including wide integers.

Type Aliases

OtherSign

Access the associated OtherSign type from an int (helper to avoid ambiguous associated types).